+++ go.mod
@@ -1,0 +1,3 @@
+module example/felipe/gowebapp
+
+go 1.19
